pyproject.toml
[tool.poetry]
name = "data_io"
version = "0.1.0"
description = "Hackable data interface layer."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ruamel-yaml = "^0.18.7"
redis = "^5.2.1"
requests = "^2.32.3"
aiohttp = "^3.11.11"
numpy = "^2.2.1"
sounddevice = "^0.5.1"
scipy = "^1.14.1"
websocket-client = "^1.8.0"
websockets = "^14.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
